Sửa mấy cái lắc nhắc của mấy cái request và service và tạo thêm 1 đống jsonSchema cho các request

+++ src/main/java/com/example/demo/service/UserService.java
@@ -216,7 +216,7 @@ public UserRespone uploadUser(MultipartFile avatar, String email) throws IOExcep
 	if (user == null) {
 		throw new AppException(ErrorCode.USER_NOT_EXISTED);
 	}
-	if (user.getAvatarUser() != null) {
+	if (user.getAvatarUser() != null && !user.getAvatarUser().isEmpty()) {
 		uploadFileService.deleteImage(user.getPublicIdAvartarUser());
 	}
 	UploadFileRespone respone = uploadFileService.uploadFile(avatar);
